\name{apply_lm}
\alias{apply_lm}
\title{Applying Linear Models}
\description{
  xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  \code{save}.
}
\usage{
  apply_lm<-function(count.matrix, repeat.counts, group)
}
\arguments{
  \item{count.matrix}{}
  \item{repeat.counts}{}
  \item{group}{}

}
\seealso{
  \code{\link{}}.
}\examples{
  ## applying linear models
  apply_lm<-function(count.matrix = count.matrix ,repeat.counts = repeat.counts, group = c(rep("T",5), rep("N",5)))
}

\keyword{file}
