\name{merge_counts}
\alias{merge_counts}
\title{Merging gene.counts and repeat.counts}
\description{
  xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  \code{save}.
}
\usage{
  merge_counts(gene.annotation, genes.expr, repeats.expr)
}
\arguments{
  \item{gene.annotation}{}
  \item{genes.expr}{}
  \item{repeats.expr}{}

}
\seealso{
  \code{\link{}}.
}\examples{
  ## summarising repeats.counts matrix
  merge_counts(gene.annotation = gene.annotation, genes.expr = gene.counts, repeats.expr = repeat.counts)
}

\keyword{file}
