\name{rm_count}
\alias{rm_count}
\title{Counting repeats the number of overlaps in each BAM file }
\description{
  xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  \code{save}.
}
\usage{
  rm_count<-function(bamlist,namelist,ranges)
}
\arguments{
  \item{bamlist}{}
  \item{namelist}{}
  \item{ranges}{}

}
\seealso{
  \code{\link{}}.
}\examples{
  ## counting repeats
  rm_count(bamlist = bamlist, namelist = namelist, ranges = ranges)
}

\keyword{file}
