\name{summarise_repeat_counts}
\alias{summarise_repeat_counts}
\title{To summarise counts of repeats as gene name and repeat family}
\description{
  xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  \code{save}.
}
\usage{
  summarise_repeat_counts(counts, namelist)
}
\arguments{
  \item{counts}{}
  \item{namelist}{}

}
\seealso{
  \code{\link{}}.
}\examples{
  ## summarising repeats.counts matrix
  summarise_repeat_counts(counts = repeats.counts, namelist = namelist)
}

\keyword{file}
