\name{summarise_rm_count}
\alias{summarise_rm_count}
\title{To summarise counts of repeats as gene name and repeat class}
\description{
  xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  \code{save}.
}
\usage{
  summarise_rm_count(counts, namelist)
}
\arguments{
  \item{counts}{}
  \item{namelist}{}

}
\seealso{
  \code{\link{}}.
}\examples{
  ## summarising repeats.counts matrix
  co_summarise(counts = repeats.counts, namelist = namelist)
}

\keyword{file}
