\name{count_repeats}
\alias{count_repeats}
\title{Counting repeats the number of overlaps in each BAM file }
\description{
  xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  \code{save}.
}
\usage{
  count_repeats(bamlist,namelist,ranges)
}
\arguments{
  \item{bamlist}{}
  \item{namelist}{}
  \item{ranges}{}

}
\seealso{
  \code{\link{}}.
}\examples{
  ## counting repeats
  count_repeats(bamlist = bamlist, namelist = namelist, ranges = ranges)
}

\keyword{file}
