\name{rm_format}
\alias{rm_format}
\title{Formatting repeat masker file}
\description{
  xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
  \code{save}.
}
\usage{
  rm_format(filepath)
}
\arguments{
  \item{filepath}{}

}
\seealso{
  \code{\link{}}.
}\examples{
  ## reading and reformatting repeat masker file
   rm<-rm_format(filepath="~/Downloads/hg38.fa.out.gz")
}

\keyword{file}
